fix(frontend): harden onboarding recording controls
